Fetch phab diff tags, if they're not available locally
Summary:
Previously, you had to do `git fetch origin --tags` before running
`git co D12345`.  Now, we'll automatically pull the tag if it exists.  I
implemented this in `revision-to-diffid.sh` rather than `git co` itself,
which is a bit of a hack, but it's a lot more covenient since that's
already a full shell script, and it knows when we are looking at a
phabricator tag at all.

Test Plan:
Find some diff ID, check that I don't have the tag locally, and then do
`git checkout D12345` on it, and it should pull the tag.

Reviewers: csilvers

Reviewed By: csilvers

Subscribers: jeremy, #devops

Differential Revision: https://phabricator.khanacademy.org/D44947