Improve searching for a branch-name in `git reflog`.
I was doing a grep for `branch-name`, but this was matching all
branches which had `branch-name` as a substring.  (Actually as a
prefix, given how the regexp turned out.)  Now I look for a more exact
match.

Auditors: benkraft

Test plan:
Ran `git reparent` when I the parent branch was user-data but I also
had done work recently on user-data-is_advocate, and it succeeded
after this change but failed before.