Don't complain about review-branches off of master outside webapp
Summary:
Outside of webapp, that's the normal workflow!  I think we had never
noticed because most folks who work outside webapp don't use it, which
is fine, but confusing, as otherwise you can just follow the normal docs
but with master as your "deploy branch".  Fixes INFRA-2283.

I did things both the good way -- depending on D50376 -- and a hackier
way as a fallback.

Test Plan:
- in this repo, `git rb foo` works just fine off master
- in webapp, it still complains
- it also complains if I do `ka-clone --repair --protect-master` first

Reviewers: csilvers, slim

Reviewed By: csilvers, slim

Subscribers: slim

Differential Revision: https://phabricator.khanacademy.org/D50375