Add deprecation warnings to these scripts (#6)
## Summary:
For the MFE project, we will want to make sure things like `git deploy-branch` support `main` as the default branch in repos where `main` _is_ the default branch.

To make that happen, we first need to bring the git-workflow repo into OLC (while preserving the git history), so that we can update the scripts and have them show up on folks' computers automatically. (This was a decision made in [this ADR](https://khanacademy.atlassian.net/wiki/spaces/ENG/pages/3283484677/ADR+769+Rethink+our+Standard+Local+Git+Setup+Between+Dotfiles+and+OLC+-+Part+2#Git-Workflow-Repo).)

Then update the versions in this repo to indicate that they're deprecated. (That's this PR!)

The only awkward thing is that this repo is open-source, so the messages might be a little odd to non-ka folks, but not really sure how to elegantly avoid that. 🤷‍♀️ 

Rough plan:
1. Bring git-workflow into OLC while preserving the git history ([PR here](https://github.com/Khan/our-lovely-cli/pull/970))
2. Remove any no-longer-needed-files from OLC ([PR here](https://github.com/Khan/our-lovely-cli/pull/971))
3. Combine the README.md in OLC ([PR here](https://github.com/Khan/our-lovely-cli/pull/972))
4. Update git-workflow repo with deprecation messages  👈 **YOU ARE HERE**
5. Update khan-dotfiles repo to no long install git-workflow ([PR here](https://github.com/Khan/khan-dotfiles/pull/142))
6. Update the scripts to support `main`, etc. ([PR here](https://github.com/Khan/our-lovely-cli/pull/973))

Issue: FEI-6455

## Test plan:
Run the scripts and see the warning. Only works if the OLC versions aren't already on your computer.

Author: lillialexis

Reviewers: csilvers

Required Reviewers:

Approved By: csilvers

Checks:

Pull Request URL: https://github.com/Khan/git-workflow/pull/6